% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{check.df}
\alias{check.df}
\title{check.df}
\usage{
check.df(df)
}
\arguments{
\item{df}{A data frame with two or three columns. Users may code positive
results as strings (positive, pos, p, yes, y, +, 1), numeric (1) or
factor (1). Users may code positive results as strings (negative, neg, no,
n, -, 0, 2), numeric (0, 2) or factor (0, 2). Case and white space
insensitive.}
}
\value{
A Boolean value indicating whether all entries in the data frame meet
the requirements described in arguments.
}
\description{
Takes a data frame with two or three columns containing the
results of binary diagnostic test(s) and a gold standard test and checks it
conforms to standards required for the compareR function.
}
\examples{
\dontrun{

# if necessary internal function calls can be made using the triple colon
# operator, but this is not shown in examples

# simulate data
df <- data.frame(
  column1 = c(" positive ", "POS ", " n ", "N ", " 1 ", "+"),
  column2 = c(" NEG ", " yes ", " negative ", " Y ", "-", " 0 ")
)

# check the dataframe
checked <- check.df(df)
checked

}

}
\keyword{internal}
