% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{ciplr}
\alias{ciplr}
\title{ciplr}
\usage{
ciplr(s1, s0, r1, r0, z)
}
\arguments{
\item{s1}{Number of cases with a positive test and positive disease status.}

\item{s0}{Number of cases with a positive test and negative disease status.}

\item{r1}{Number of cases with a negative test and positive disease status.}

\item{r0}{Number of cases with a negative test and negative disease status.}

\item{z}{Z-score of normal distribution at a defined alpha level.}
}
\value{
A list object of class `'conf.1t'` or `'conf.2t'` containing lower
and upper confidence bounds for positive likelihood ratio.
}
\description{
Calculates confidence intervals for positive likelihood ratio.
}
\examples{
\dontrun{

# if necessary internal function calls can be made using the triple colon
# operator, but this is not shown in examples

# load data
df <- cfpr

# drop first column
df <- subset(df, select = 2:3)

# compute important values
vals <- values.1test(df)

# set value of z
z <- qnorm(1 - 0.05 / 2)

# compute confidence values for positive likelihood ratio
ciplr(vals$s1, vals$s0, vals$r1, vals$r0, z)

}

}
\keyword{internal}
