% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{conf.acc}
\alias{conf.acc}
\title{conf.acc}
\usage{
conf.acc(vals, alpha = 0.05, ...)
}
\arguments{
\item{vals}{A list object containing all values required for calculations with class `vals.1test` or `vals.2test`. These objects are output by the `values.1test()` and `values.2test()` function.}

\item{alpha}{An alpha value. Defaults to 0.05.}

\item{...}{Rarely needs to be used. Allows additional arguments to be passed to internal functions.}
}
\value{
A list object of class `'conf.1t'` or `'conf.2t'` containing point estimates, plus lower and upper confidence bounds for diagnostic accuracies.
}
\description{
Calculates confidence intervals for accuracies (sensitivity and specificity) for one or two binary diagnostic tests.
}
\examples{
# load data
df <- cfpr

# compute important values
vals <- testCompareR:::values.2test(df)

# compute confidence intervals for diagnostic accuracies
testCompareR:::conf.acc(vals, alpha = 0.05)

}
\keyword{internal}
