% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeR.R
\name{dataframeR}
\alias{dataframeR}
\title{dataframeR}
\usage{
dataframeR(s11, s10, s01, s00, r11, r10, r01, r00)
}
\arguments{
\item{s11}{Number of cases where Test 1 is positive, Test 2 is positive and
gold standard is positive.}

\item{s10}{Number of cases where Test 1 is positive, Test 2 is negative and
gold standard is positive.}

\item{s01}{Number of cases where Test 1 is negative, Test 2 is positive and
gold standard is positive.}

\item{s00}{Number of cases where Test 1 is negative, Test 2 is negative and
gold standard is positive.}

\item{r11}{Number of cases where Test 1 is positive, Test 2 is positive and
gold standard is negative.}

\item{r10}{Number of cases where Test 1 is positive, Test 2 is negative and
gold standard is negative.}

\item{r01}{Number of cases where Test 1 is negative, Test 2 is positive and
gold standard is negative.}

\item{r00}{Number of cases where Test 1 is negative, Test 2 is negative and
gold standard is negative.}
}
\value{
A data frame populated with zeros and ones indicating positive or
negative test results which can be passed to the compareR function.
}
\description{
Produces a data frame which can be used by the compareR function
using values commonly found in published literature. Useful for reviews and
meta-analyses.
}
\details{
Understanding the parameter names:
s & r represent positive and negative results for the gold standard test,
respectively.
The first digit represents a positive (1) or negative (0) result for Test 1.
The second digit represents a positive (1) or negative (0) result for Test 2.
}
\examples{
# build data frame using numbers
dataframeR(3, 3, 3, 3, 3, 3, 3, 3)

}
