% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{disp.cont}
\alias{disp.cont}
\title{disp.cont}
\usage{
disp.cont(vals, margins = FALSE, ...)
}
\arguments{
\item{vals}{A list object containing all values required for calculations with class `'vals.1test'` or `'vals.2test'`. These objects are output by the `values.1test()` and `values.2test()` function.}

\item{margins}{A Boolean value indicating whether the outputted matrix should have margins containing summed totals of rows and columns.}

\item{...}{Rarely needs to be used. Allows additional arguments to be passed to internal functions.}
}
\value{
Either a matrix or list of matrices displaying counts of positive and negative tests by true disease status.
}
\description{
Creates and displays a contingency table of positive and negative test results and disease status.
}
\examples{
# load data
df <- cfpr

# compute important values
vals <- testCompareR:::values.2test(df)

# compute contingency table
testCompareR:::disp.cont(vals, margins = TRUE)

}
\keyword{internal}
