% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretR.R
\name{interpretR}
\alias{interpretR}
\title{interpretR}
\usage{
interpretR(result)
}
\arguments{
\item{result}{A list object with class 'compareR' output from the compareR
function.}
}
\value{
A plain English summary of the findings produced by the compareR
function.
}
\description{
Provides a plain English readout of the results of the compareR
function.
}
\examples{
# simulate data
test1 <- c(rep(1, 300), rep(0, 100), rep(1, 55), rep(0, 145))
test2 <- c(rep(1, 280), rep(0, 120), rep(1, 45), rep(0, 155))
gold <- c(rep(1, 400), rep(0, 200))
dat <- data.frame(test1, test2, gold)

# compare with compareR
result <- compareR(dat)

# provide a plain English readout with interpretR
interpretR(result)

}
