% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{matrixify}
\alias{matrixify}
\title{matrixify}
\usage{
matrixify(
  vals,
  rows = c("Row 1", "Row 2"),
  test.names = c("Test 1", "Test 2"),
  dp = 3
)
}
\arguments{
\item{vals}{A list object containing all values required for calculations with class `vals.1test` or `vals.2test`. These objects are output by the `values.1test()` and `values.2test()` function.}

\item{rows}{A vector of length two giving the names of the related tests (eg. sensitivity and specificity).}

\item{test.names}{A vector of length two giving the names of the two different binary diagnostic tests. This argument is not relevant when testing a single binary diagnostic test.}

\item{dp}{Number of decimal places of output. Defaults to 1.}
}
\value{
Either a matrix or list of matrices displaying point estimates, standard errors and upper and lower confidence bounds for related outputs (eg. sensitivity and specificity) of one or two binary diagnostic tests.
}
\description{
Produces a matrix containing point estimates, standard errors and upper and lower confidence bounds for related outputs (eg. sensitivity and specificity) of one or two binary diagnostic tests.
}
\examples{
# load data
df <- cfpr

# compute important values
vals <- testCompareR:::values.2test(df)

# compute confidence for diagnostic accuracies
acc <- testCompareR:::conf.acc(vals)

# produce a matrix summarising results
testCompareR:::matrixify(acc)

}
\keyword{internal}
