% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.compareR}
\alias{plot.compareR}
\title{Plot a compareR object}
\usage{
\method{plot}{compareR}(x, ...)
}
\arguments{
\item{x}{An object of class compareR.}

\item{...}{Arguments such as graphical parameters. Not currently in use.}
}
\value{
A visualisation of the results for diagnostic accuracies and
predictive values from the compareR output.
}
\description{
An S3 method to plot a simple visualisation of the results from
the compareR function.
}
\details{
Method to plot the most commonly used results of the compareR output.
}
\examples{
# generate result
res <- compareR(cass, test1 = "exercise", test2 = "cp",
                gold = "angio",
                test.names = c("ExerciseStressTest", "ChestPain"))

# run print method
plot(res)

}
