% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.compareR}
\alias{print.compareR}
\title{Print a compareR object}
\usage{
\method{print}{compareR}(x, ...)
}
\arguments{
\item{x}{An object of class compareR.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A printed results table of the compareR output.
}
\description{
An S3 method to print the results verbose from the
compareR function.
}
\details{
Method to print the pertinent results of the compareR output.
}
\examples{
# generate result
res <- compareR(cass, test1 = "exercise", test2 = "cp",
                gold = "angio",
                test.names = c("ExerciseStressTest", "ChestPain"))

# run print method
print(res)

}
