% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{recoder}
\alias{recoder}
\title{recoder}
\usage{
recoder(df)
}
\arguments{
\item{df}{A dataframe with two or three columns. Users may code positive results as strings (positive, pos, p, yes, y, +, 1), numeric (1) or factor (1). Users may code positive results as strings (negative, neg, no, n, -, 0, 2), numeric (0, 2) or factor (0, 2). Case and white space insensitive.}
}
\value{
A data frame with the same number of columns as df re-coded to integers of 0 (negative) and 1 (positive).
}
\description{
Takes a data frame with two or three columns containing the results of binary diagnostic test(s) and a gold standard test and re-codes to an acceptable format for the compareR function.
}
\examples{
# simulate data
df <- data.frame(
  column1 = c(" positive ", "POS ", " n ", "N ", " 1 ", "+"),
  column2 = c(" NEG ", " yes ", " negative ", " Y ", "-", " 0 ")
)

# recode the dataframe
recoded_df <- testCompareR:::recoder(df)
recoded_df

}
\keyword{internal}
