% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.compareR}
\alias{summary.compareR}
\title{Summarise a compareR object}
\usage{
\method{summary}{compareR}(object, ...)
}
\arguments{
\item{object}{An object of class compareR.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
A summary of the compareR output.
}
\description{
An S3 method to summarise the rather verbose output of the
compareR function.
}
\details{
Method to summarise the verbose compareR output.
}
\examples{
# generate result
res <- compareR(cass, test1 = "exercise", test2 = "cp",
                gold = "angio",
                test.names = c("ExerciseStressTest", "ChestPain"))

# run summary method
summary(res)

}
