% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{values.1test}
\alias{values.1test}
\title{values.1test}
\usage{
values.1test(df)
}
\arguments{
\item{df}{A dataframe with 2 columns (test1, gold), populated with zeros and ones.}
}
\value{
A list object containing important values for future calculations.
}
\description{
Creates a list object containing important values for future calculations.
}
\examples{
# load data
df <- cfpr

# drop first column
df <- subset(df, select = 2:3)

# compute important values
testCompareR:::values.1test(df)

}
\keyword{internal}
