% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{values.2test}
\alias{values.2test}
\title{values.2test}
\usage{
values.2test(df)
}
\arguments{
\item{df}{A dataframe with 3 columns (test1, test2, gold), populated with zeros and ones.}
}
\value{
A list object containing important values for future calculations.
}
\description{
Creates a list object containing important values for future calculations.
}
\examples{
# load data
df <- cfpr

# compute important values
testCompareR:::values.2test(df)

}
\keyword{internal}
