% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareR_internal.R
\name{yu.int}
\alias{yu.int}
\title{yu.int}
\usage{
yu.int(n, z, est)
}
\arguments{
\item{n}{Total number of cases evaluated.}

\item{z}{Z-score of normal distribution at a defined alpha level.}

\item{est}{The point estimate of the specific test under evaluation.}

\item{...}{Rarely needs to be used. Allows additional arguments to be passed
to internal functions.}
}
\value{
A list object containing lower and upper confidence bounds.
}
\description{
Calculates the Yu et al. confidence interval for binomial
proportions.
}
\examples{
\dontrun{

# if necessary internal function calls can be made using the triple colon
# operator, but this is not shown in examples

# load data
df <- cfpr

# compute important values
vals <- values.2test(df)

# set value of z
z <- qnorm(1 - 0.05 / 2)

# compute Yu interval for disease prevalence
yu.int(vals$n, z, vals$prev)

}

}
\keyword{internal}
